Module mongodb::action

Action builder types.

Modules

gridfs

Action builders for gridfs.

Structs

AbortTransaction

Abort the currently active transaction on a session. Construct with ClientSession::abort_transaction.

Aggregate

Run an aggregation operation. Construct with Database::aggregate or Collection::aggregate.

BulkWrite

Performs multiple write operations. Construct with Client::bulk_write.

CommitTransaction

Commits a currently-active transaction. Construct with ClientSession::commit_transaction.

CountDocuments

Get an accurate count of documents. Construct with Collection::count_documents.

CreateCollection

Creates a new collection. Construct with Database::create_collection.

CreateIndex

Perform creation of an index or indexes. Construct by calling Collection::create_index or Collection::create_indexes.

CreateSearchIndex

Create search indexes on a collection. Construct with Collection::create_search_index or Collection::create_search_indexes.

Delete

Deletes documents matching a query. Construct with Collection::delete_one or Collection::delete_many.

Distinct

Finds the distinct values of a field. Construct with Collection::distinct.

DropCollection

Drops the collection, deleting all data and indexes stored in it. Construct with Collection::drop.

DropDatabase

Drops the database, deleting all data, collections, and indexes stored in it. Construct with Database::drop.

DropIndex

Drop an index or indexes. Construct with Collection::drop_index or Collection::drop_indexes.

DropSearchIndex

Drops a specific search index. Construct with Collection::drop_search_index.

EstimatedDocumentCount

Gather an estimated document count. Construct with Collection::estimated_document_count.

ExplicitSession

Find

Finds the documents in a collection matching a filter. Construct with Collection::find.

FindOne

Finds a single document in a collection matching a filter. Construct with Collection::find_one.

FindOneAndDelete

Atomically finds up to one document in the collection matching a filter and deletes it. Construct with Collection::find_one_and_delete.

FindOneAndReplace

Atomically finds up to one document in the collection matching a filter and replaces it. Construct with Collection::find_one_and_replace.

FindOneAndUpdate

Atomically finds up to one document in the collection matching a filter and updates it. Construct with Collection::find_one_and_update.

ImplicitSession

InsertMany

Inserts documents into a collection. Construct with Collection::insert_many.

InsertOne

Inserts a document into a collection. Construct with [’Collection::insert_one`].

ListCollections

Gets information about each of the collections in the database. Create by calling Database::list_collections or Database::list_collection_names.

ListDatabases

Gets information about each database present in the cluster the Client is connected to. Create by calling Client::list_databases or Client::list_database_names.

ListIndexes

List indexes on a collection. Construct with Collection::list_indexes or Collection::list_index_names.

ListNames

ListSearchIndexes

Gets index information for one or more search indexes in a collection.

ListSpecifications

Multiple

ParseConnectionString

Parses a MongoDB connection string into a ClientOptions struct. Construct with ClientOptions::parse.

ReplaceOne

Replace up to one document matching a query. Construct with Collection::replace_one.

RunCommand

Run a database-level command. Create with Database::run_command.

RunCursorCommand

Runs a database-level command and returns a cursor to the response. Create with Database::run_cursor_command.

Shutdown

Shut down this Client, terminating background thread workers and closing connections. Create by calling Client::shutdown.

Single

StartSession

Starts a new ClientSession. Construct with Client::start_session.

StartTransaction

Start a new transaction. Construct with ClientSession::start_transaction.

Update

Update documents matching a query. Construct with Collection::update_many or Collection::update_one.

UpdateSearchIndex

Updates a specific search index to use a new definition. Construct with Collection::update_search_index.

WarmConnectionPool

Add connections to the connection pool up to min_pool_size. Construct with Client::warm_connection_pool.

Watch

Starts a new ChangeStream that receives events for all changes in a given scope. Create by calling Client::watch, Database::watch, or Collection::watch.

Traits

Action

A pending action to execute on the server. The action can be configured via chained methods and executed via await (or run if using the sync client).