Struct mongodb::db::DatabaseInner

pub struct DatabaseInner {
    pub name: String,
    pub client: Client,
    pub read_preference: ReadPreference,
    pub write_concern: WriteConcern,
}

Interfaces with a MongoDB database.

Fields

name

The database name.

client

A reference to the client that spawned this database.

read_preference

Indicates how a server should be selected for read operations.

write_concern

Describes the guarantees provided by MongoDB when reporting the success of a write operation.