Enum mongodb::error::ErrorKind

pub enum ErrorKind {
    AddrParse(AddrParseError),
    ArgumentError {
        message: String,
    },
    AuthenticationError {
        message: String,
    },
    BsonDecode(DecoderError),
    BsonEncode(EncoderError),
    BulkWriteError(BulkWriteFailure),
    CommandError(CommandError),
    DnsResolve(ResolveError),
    InvalidDnsName(InvalidDNSNameError),
    InvalidHostname {
        hostname: String,
    },
    Io(Error),
    NoDnsResults(StreamAddress),
    OperationError {
        message: String,
    },
    OutOfRangeError(OutOfRangeError),
    ParseError {
        data_type: String,
        file_path: String,
    },
    ResponseError {
        message: String,
    },
    ServerSelectionError {
        message: String,
    },
    SrvLookupError {
        message: String,
    },
    TxtLookupError {
        message: String,
    },
    WaitQueueTimeoutError {
        address: StreamAddress,
    },
    WriteError(WriteFailure),
}

The types of errors that can occur.

Variants

AddrParse(AddrParseError)

Wrapper around std::net::AddrParseError.

ArgumentError

An invalid argument was provided to a database operation.

Fields of ArgumentError

message: String

AuthenticationError

An error occurred while the Client attempted to authenticate a connection.

Fields of AuthenticationError

message: String

BsonDecode(DecoderError)

Wrapper around bson::DecoderError.

BsonEncode(EncoderError)

Wrapper around bson::EncoderError.

BulkWriteError(BulkWriteFailure)

An error occurred when trying to execute a write operation consisting of multiple writes.

CommandError(CommandError)

The server returned an error to an attempted operation.

DnsResolve(ResolveError)

Wrapper around trust_dns_resolver::error::ResolveError.

InvalidDnsName(InvalidDNSNameError)

Wrapper around webpki::InvalidDNSNameError.

InvalidHostname

A hostname could not be parsed.

Fields of InvalidHostname

hostname: String

Io(Error)

Wrapper around std::io::Error.

NoDnsResults(StreamAddress)

OperationError

A database operation failed to send or receive a reply.

Fields of OperationError

message: String

OutOfRangeError(OutOfRangeError)

ParseError

Data from a file could not be parsed.

Fields of ParseError

data_type: Stringfile_path: String

ResponseError

The server returned an invalid reply to a database operation.

Fields of ResponseError

message: String

ServerSelectionError

The Client was not able to select a server for the operation.

Fields of ServerSelectionError

message: String

SrvLookupError

An error occurred during SRV record lookup.

Fields of SrvLookupError

message: String

TxtLookupError

An error occurred during TXT record lookup

Fields of TxtLookupError

message: String

WaitQueueTimeoutError

The Client timed out while checking out a connection from connection pool.

Fields of WaitQueueTimeoutError

address: StreamAddress

WriteError(WriteFailure)

An error occurred when trying to execute a write operation