mongodb
Module event

Contains the events and functionality for monitoring internal Client behavior.

Modules

cmap

Contains the events and functionality for monitoring behavior of the connection pooling of a Client.

command

Contains the events and functionality to monitor the commands and responses that a Client sends and receives from the server.

sdam

Contains the events and functionality for monitoring Server Discovery and Monitoring.

Enums

EventHandler

A destination for events. Allows implicit conversion via From for concrete types for convenience with crate::options::ClientOptions construction: