mongodb::event
Module cmap

Contains the events and functionality for monitoring behavior of the connection pooling of a Client.

Structs

ConnectionCheckedInEvent

Event emitted when a connection is checked back into a connection pool.

ConnectionCheckedOutEvent

Event emitted when a connection is successfully checked out.

ConnectionCheckoutFailedEvent

Event emitted when a thread is unable to check out a connection.

ConnectionCheckoutStartedEvent

Event emitted when a thread begins checking out a connection to use for an operation.

ConnectionClosedEvent

Event emitted when a connection is closed.

ConnectionCreatedEvent

Event emitted when a connection is created.

ConnectionPoolOptions

Contains the options for creating a connection pool. While these options are specified at the client-level, ConnectionPoolOptions is exposed for the purpose of CMAP event handling.

ConnectionReadyEvent

Event emitted when a connection is ready to be used. This indicates that all the necessary prerequisites for using a connection (handshake, authentication, etc.) have been completed.

PoolClearedEvent

Event emitted when a connection pool is cleared.

PoolClosedEvent

Event emitted when a connection pool is cleared.

PoolCreatedEvent

Event emitted when a connection pool is created.

PoolReadyEvent

Event emitted when a connection pool becomes ready.

Enums

CmapEvent

ConnectionCheckoutFailedReason

The reasons a connection may not be able to be checked out.

ConnectionClosedReason

The reasons that a connection may be closed.

Traits

CmapEventHandlerDeprecated

Usage of this trait is deprecated. Applications should use the EventHandler API.