Struct mongodb::event::cmap::ConnectionPoolOptions

pub struct ConnectionPoolOptions {
    pub app_name: Option<String>,
    pub connect_timeout: Option<Duration>,
    pub credential: Option<Credential>,
    pub event_handler: Option<Arc<dyn CmapEventHandler>>,
    pub max_idle_time: Option<Duration>,
    pub max_pool_size: Option<u32>,
    pub min_pool_size: Option<u32>,
    pub tls_options: Option<TlsOptions>,
    pub wait_queue_timeout: Option<Duration>,
}

Contains the options for creating a connection pool. While these options are specified at the client-level, ConnectionPoolOptions is exposed for the purpose of CMAP event handling.

Fields
app_name: Option<String>

The application name specified by the user. This is sent to the server as part of the handshake that each connection makes when it's created.

connect_timeout: Option<Duration>

The connect timeout passed to each underlying TcpStream when attemtping to connect to the server.

credential: Option<Credential>

The credential to use for authenticating connections in this pool.

event_handler: Option<Arc<dyn CmapEventHandler>>

Processes all events generated by the pool.

max_idle_time: Option<Duration>

Connections that have been ready for usage in the pool for longer than max_idle_time will not be used.

The default is that connections will not be closed due to being idle.

max_pool_size: Option<u32>

The maximum number of connections that the pool can have at a given time. This includes connections which are currently checked out of the pool.

The default is 100.

min_pool_size: Option<u32>

The minimum number of connections that the pool can have at a given time. This includes connections which are currently checked out of the pool. If fewer than min_pool_size connections are in the pool, connections will be added to the pool in the background.

The default is that no minimum is enforced

tls_options: Option<TlsOptions>

The options specifying how a TLS connection should be configured. If tls_options is None, then TLS will not be used for the connections.

The default is not to use TLS for connections.

wait_queue_timeout: Option<Duration>

Rather than wait indefinitely for a connection to become available, instead return an error after the given duration.

The default is to block indefinitely until a connection becomes available.

Methods

impl ConnectionPoolOptions

pub fn builder(
) -> TypedBuilder_BuilderFor_ConnectionPoolOptions<(), (), (), (), (), (), (), (), ()>

Create a builder for building ConnectionPoolOptions. On the builder, call .app_name(...)(optional), .connect_timeout(...)(optional), .credential(...)(optional), .event_handler(...)(optional), .max_idle_time(...)(optional), .max_pool_size(...)(optional), .min_pool_size(...)(optional), .tls_options(...)(optional), .wait_queue_timeout(...)(optional) to set the values of the fields(they accept Into values). Finally, call .build() to create the instance of ConnectionPoolOptions.