mongodb
Module gridfs

Contains the functionality for GridFS operations.

Structs

FilesCollectionDocument

A model for the documents stored in a GridFS bucket’s files collection.

GridFsBucket

A GridFsBucket provides the functionality for storing and retrieving binary BSON data that exceeds the 16 MiB size limit of a MongoDB document. Users may upload and download large amounts of data, called files, to the bucket. When a file is uploaded, its contents are divided into chunks and stored in a chunks collection. A corresponding FilesCollectionDocument is also stored in a files collection. When a user downloads a file, the bucket finds and returns the data stored in its chunks.

GridFsDownloadStream

A stream from which a file stored in a GridFS bucket can be downloaded.

GridFsUploadStream

A stream to which bytes can be written to be uploaded to a GridFS bucket.