Struct mongodb::gridfs::file::File

pub struct File {
    pub doc: GfsFile,
    // some fields omitted
}

A writable or readable file stream within GridFS.

Fields
doc: GfsFile

The file document associated with this stream.

Methods

impl File

fn new(gfs: Store, id: ObjectId, mode: Mode) -> File

A new file stream with an id-referenced GridFS file.

fn with_name(gfs: Store, name: String, id: ObjectId, mode: Mode) -> File

A new file stream with a name-and-id-referenced GridFS file.

fn with_doc(gfs: Store, doc: Document) -> File

A new file stream from a read file document.

fn len(&self) -> i64

Returns the byte length of the file.

fn is_empty(&self) -> bool

Returns true if the file contains no bytes.

fn err_description(&self) -> Result<Option<String>>

Retrieves the description of the threaded error, if one occurred.

fn assert_mode(&self, mode: Mode) -> Result<()>

Ensures the file mode matches the desired mode.

fn close(&mut self) -> Result<()>

Completes writing or reading and closes the file. This will be called when the file is dropped, but errors will be ignored. Therefore, this method should be called manually.

fn find_chunk(&mut self, id: ObjectId, chunk_num: i32) -> Result<Vec<u8>>