Struct mongodb::gridfs::GridFsBucket

pub struct GridFsBucket { /* private fields */ }

A GridFsBucket provides the functionality for storing and retrieving binary BSON data that exceeds the 16 MiB size limit of a MongoDB document. Users may upload and download large amounts of data, called files, to the bucket. When a file is uploaded, its contents are divided into chunks and stored in a chunks collection. A corresponding FilesCollectionDocument is also stored in a files collection. When a user downloads a file, the bucket finds and returns the data stored in its chunks.

GridFsBucket uses std::sync::Arc internally, so it can be shared safely across threads or async tasks.

Implementations

impl GridFsBucket

pub async fn download_to_futures_0_3_writer<T>( &self, id: Bson, destination: T ) -> Result<()>where T: AsyncWrite + Unpin,

Downloads the contents of the stored file specified by id and writes the contents to the destination, which may be any type that implements the futures_io::AsyncWrite trait.

To download to a type that implements tokio::io::AsyncWrite, use the tokio_util::compat module to convert between types.

use tokio_util::compat::TokioAsyncWriteCompatExt;

let futures_writer = tokio_writer.compat_write();
bucket.download_to_futures_0_3_writer(id, futures_writer).await?;

Note that once an AsyncWrite trait is stabilized in the standard library, this method will be deprecated in favor of one that accepts a std::io::AsyncWrite source.

pub async fn download_to_futures_0_3_writer_by_name<T>( &self, filename: impl AsRef<str>, destination: T, options: impl Into<Option<GridFsDownloadByNameOptions>> ) -> Result<()>where T: AsyncWrite + Unpin,

Downloads the contents of the stored file specified by filename and writes the contents to the destination, which may be any type that implements the futures_io::AsyncWrite trait.

If there are multiple files in the bucket with the given filename, the revision in the options provided is used to determine which one to download. See the documentation for GridFsDownloadByNameOptions for details on how to specify a revision. If no revision is provided, the file with filename most recently uploaded will be downloaded.

To download to a type that implements tokio::io::AsyncWrite, use the tokio_util::compat module to convert between types.

use tokio_util::compat::TokioAsyncWriteCompatExt;

let futures_writer = tokio_writer.compat_write();
bucket.download_to_futures_0_3_writer_by_name("example_file", futures_writer, None).await?;

Note that once an AsyncWrite trait is stabilized in the standard library, this method will be deprecated in favor of one that accepts a std::io::AsyncWrite source.

impl GridFsBucket

pub async fn open_download_stream( &self, id: Bson ) -> Result<GridFsDownloadStream>

Opens and returns a GridFsDownloadStream from which the application can read the contents of the stored file specified by id.

pub async fn open_download_stream_by_name( &self, filename: impl AsRef<str>, options: impl Into<Option<GridFsDownloadByNameOptions>> ) -> Result<GridFsDownloadStream>

Opens and returns a GridFsDownloadStream from which the application can read the contents of the stored file specified by filename.

If there are multiple files in the bucket with the given filename, the revision in the options provided is used to determine which one to download. See the documentation for GridFsDownloadByNameOptions for details on how to specify a revision. If no revision is provided, the file with filename most recently uploaded will be downloaded.

impl GridFsBucket

pub async fn upload_from_futures_0_3_reader<T>( &self, filename: impl AsRef<str>, source: T, options: impl Into<Option<GridFsUploadOptions>> ) -> Result<ObjectId>where T: AsyncRead + Unpin,

Uploads a user file to the bucket. Bytes are read from source, which may be any type that implements the futures_io::AsyncRead trait, and stored in chunks in the bucket’s chunks collection. After all the chunks have been uploaded, a corresponding FilesCollectionDocument is stored in the bucket’s files collection.

This method generates an ObjectId for the id field of the FilesCollectionDocument and returns it.

To upload from a type that implements tokio::io::AsyncRead, use the tokio_util::compat module to convert between types.

use tokio_util::compat::TokioAsyncReadCompatExt;

let futures_reader = tokio_reader.compat();
bucket.upload_from_futures_0_3_reader("example_file", futures_reader, None).await?;

Note that once an AsyncRead trait is stabilized in the standard library, this method will be deprecated in favor of one that accepts a std::io::AsyncRead source.

pub async fn upload_from_futures_0_3_reader_with_id<T>( &self, id: Bson, filename: impl AsRef<str>, source: T, options: impl Into<Option<GridFsUploadOptions>> ) -> Result<()>where T: AsyncRead + Unpin,

Uploads a user file to the bucket. Bytes are read from source, which may be any type that implements the futures_io::AsyncRead trait, and stored in chunks in the bucket’s chunks collection. After all the chunks have been uploaded, a corresponding FilesCollectionDocument with the given id is stored in the bucket’s files collection.

To upload from a type that implements tokio::io::AsyncRead, use the tokio_util::compat module to convert between types.

use tokio_util::compat::TokioAsyncReadCompatExt;

let futures_reader = tokio_reader.compat();
bucket.upload_from_futures_0_3_reader_with_id(id, "example_file", futures_reader, None).await?;

Note that once an AsyncRead trait is stabilized in the standard library, this method will be deprecated in favor of one that accepts a std::io::AsyncRead source.

impl GridFsBucket

pub fn open_upload_stream( &self, filename: impl AsRef<str>, options: impl Into<Option<GridFsUploadOptions>> ) -> GridFsUploadStream

Creates and returns a GridFsUploadStream that the application can write the contents of the file to. This method generates a unique ObjectId for the corresponding FilesCollectionDocument’s id field that can be accessed via the stream’s id method.

pub fn open_upload_stream_with_id( &self, id: Bson, filename: impl AsRef<str>, options: impl Into<Option<GridFsUploadOptions>> ) -> GridFsUploadStream

Opens a GridFsUploadStream that the application can write the contents of the file to. The provided id will be used for the corresponding FilesCollectionDocument’s id field.

impl GridFsBucket

pub fn read_concern(&self) -> Option<&ReadConcern>

Gets the read concern of the bucket.

pub fn write_concern(&self) -> Option<&WriteConcern>

Gets the write concern of the bucket.

pub fn selection_criteria(&self) -> Option<&SelectionCriteria>

Gets the selection criteria of the bucket.

pub async fn delete(&self, id: Bson) -> Result<()>

Deletes the FilesCollectionDocument with the given id and its associated chunks from this bucket. This method returns an error if the id does not match any files in the bucket.

pub async fn find( &self, filter: Document, options: impl Into<Option<GridFsFindOptions>> ) -> Result<Cursor<FilesCollectionDocument>>

Finds and returns the FilesCollectionDocuments within this bucket that match the given filter.

pub async fn rename( &self, id: Bson, new_filename: impl AsRef<str> ) -> Result<()>

Renames the file with the given ‘id’ to the provided new_filename. This method returns an error if the id does not match any files in the bucket.

pub async fn drop(&self) -> Result<()>

Removes all of the files and their associated chunks from this bucket.