Struct mongodb::gridfs::GridFsDownloadStream

pub struct GridFsDownloadStream { /* private fields */ }

A stream from which a file stored in a GridFS bucket can be downloaded.

Downloading from the Stream

The GridFsDownloadStream type implements futures_io::AsyncRead. It is recommended that users call the utility methods in AsyncReadExt to interact with the stream.

use futures_util::io::AsyncReadExt;

let mut buf = Vec::new();
let mut download_stream = bucket.open_download_stream(id).await?;
download_stream.read_to_end(&mut buf).await?;

Using tokio::io::AsyncRead

Users who prefer to use tokio’s AsyncRead trait can use the tokio_util::compat module.

use tokio_util::compat::FuturesAsyncReadCompatExt;

let futures_upload_stream = bucket.open_download_stream(id).await?;
let tokio_upload_stream = futures_upload_stream.compat();