mongodb::options
Module oidc

Contains the functionality for OIDC authorization and authentication. Contains the functionality for OIDC authorization and authentication.

Structs

Callback

The callback to use for OIDC authentication.

CallbackContext

CallbackContext contains the information necessary to perform the human or machine flow in a function. The driver passes ownership of this struct to the Callback function.

IdpServerInfo

IdpServerInfo contains the information necessary to locate and authorize with an OIDC server.

IdpServerResponse

The return type of the OIDC authentication function. It contains the access token with optional expiration time and refresh token.