Struct mongodb::options::FindOneAndReplaceOptions

pub struct FindOneAndReplaceOptions {
    pub bypass_document_validation: Option<bool>,
    pub max_time: Option<Duration>,
    pub projection: Option<Document>,
    pub return_document: Option<ReturnDocument>,
    pub sort: Option<Document>,
    pub upsert: Option<bool>,
    pub write_concern: Option<WriteConcern>,
    pub collation: Option<Collation>,
}

Specifies the options to a Collection::find_one_and_replace operation.

Fields
bypass_document_validation: Option<bool>

Opt out of document-level validation.

max_time: Option<Duration>

The maximum amount of time to allow the query to run.

This options maps to the maxTimeMS MongoDB query option, so the duration will be sent across the wire as an integer number of milliseconds.

projection: Option<Document>

Limits the fields of the document being returned.

return_document: Option<ReturnDocument>

Whether the operation should return the document before or after modification.

sort: Option<Document>

The order of the documents for the purposes of the operation.

upsert: Option<bool>

If true, insert a document if no matching document is found.

write_concern: Option<WriteConcern>

The level of the write concern

collation: Option<Collation>

The collation to use for the operation.

See the documentation for more information on how to use this option.

Methods

impl FindOneAndReplaceOptions

pub fn builder(
) -> TypedBuilder_BuilderFor_FindOneAndReplaceOptions<(), (), (), (), (), (), (), ()>

Create a builder for building FindOneAndReplaceOptions. On the builder, call .bypass_document_validation(...)(optional), .max_time(...)(optional), .projection(...)(optional), .return_document(...)(optional), .sort(...)(optional), .upsert(...)(optional), .write_concern(...)(optional), .collation(...)(optional) to set the values of the fields(they accept Into values). Finally, call .build() to create the instance of FindOneAndReplaceOptions.