Struct mongodb::ClientInner

pub struct ClientInner {
    pub read_preference: ReadPreference,
    pub write_concern: WriteConcern,
    // some fields omitted
}

Interfaces with a MongoDB server or replica set.

Fields

read_preference

Indicates how a server should be selected for read operations.

write_concern

Describes the guarantees provided by MongoDB when reporting the success of a write operation.