Struct mongodb::ClientOptions

pub struct ClientOptions {
    pub log_file: Option<String>,
    pub read_preference: Option<ReadPreference>,
    pub write_concern: Option<WriteConcern>,
    pub heartbeat_frequency_ms: u32,
    pub server_selection_timeout_ms: i64,
    pub local_threshold_ms: i64,
}

Configuration options for a client.

Fields

log_file

File path for command logging.

read_preference

Client-level server selection preferences for read operations.

write_concern

Client-level write guarantees when reporting a write success.

heartbeat_frequency_ms

Frequency of server monitor updates; default 10000 ms.

server_selection_timeout_ms

Timeout for selecting an appropriate server for operations; default 30000 ms.

local_threshold_ms

The size of the latency window for selecting suitable servers; default 15 ms.

Methods

impl ClientOptions

fn new() -> ClientOptions

Creates a new default options struct.

fn with_log_file(file: &str) -> ClientOptions

Creates a new options struct with a specified log file.