mongodb
Struct Collection

pub struct Collection<T>
where
    T: Send + Sync,
{ /* private fields */ }

Collection is the client-side abstraction of a MongoDB Collection. It can be used to perform collection-level operations such as CRUD operations. A Collection can be obtained through a Database by calling either Database::collection or Database::collection_with_options.

A Collection can be parameterized with any type that implements the Serialize and Deserialize traits from the serde crate. This includes but is not limited to just Document. The various methods that accept or return instances of the documents in the collection will accept/return instances of the generic parameter (e.g. Collection::insert_one accepts it as an argument, Collection::find_one returns an Option of it). It is recommended to define types that model your data which you can parameterize your Collections with instead of Document, since doing so eliminates a lot of boilerplate deserialization code and is often more performant.

Collection uses std::sync::Arc internally, so it can safely be shared across threads or async tasks.

Example

use serde::{Deserialize, Serialize};

/// Define a type that models our data.
#[derive(Clone, Debug, Deserialize, Serialize)]
struct Item {
    id: u32,
}

// Parameterize our collection with the model.
let coll = client.database("items").collection::<Item>("in_stock");

for i in 0..5 {
    let coll_ref = coll.clone();

    // Spawn several tasks that operate on the same collection concurrently.
    tokio::task::spawn(async move {
        // Perform operations with `coll_ref` that work with directly our model.
        coll_ref.insert_one(Item { id: i }).await;
    });
}

Implementations

impl<T> Collection<T>
where T: Send + Sync,

pub fn aggregate( &self, pipeline: impl IntoIterator<Item = Document>, ) -> Aggregate<'_>

Runs an aggregation operation.

See the documentation here for more information on aggregations.

await will return Result<Cursor<Document>>. If a ClientSession was provided, the returned cursor will be a SessionCursor. If with_type was called, the returned cursor will be generic over the T specified.

These methods can be chained before .await to set options:

with_type

with_options

allow_disk_use

batch_size

bypass_document_validation

collation

comment

hint

max_await_time

max_time

read_concern

selection_criteria

write_concern

let_vars

session

impl<T> Collection<T>
where T: Send + Sync,

pub fn estimated_document_count(&self) -> EstimatedDocumentCount<'_>

Estimates the number of documents in the collection using collection metadata.

Due to an oversight in versions 5.0.0 - 5.0.7 of MongoDB, the count server command, which estimatedDocumentCount uses in its implementation, was not included in v1 of the Stable API. Users of the Stable API with estimatedDocumentCount are recommended to upgrade their cluster to 5.0.8+ or set ServerApi::strict to false to avoid encountering errors.

For more information on the behavior of the count server command, see Count: Behavior.

await will return Result<u64>.

These methods can be chained before .await to set options:

with_options

max_time

selection_criteria

read_concern

comment

pub fn count_documents(&self, filter: Document) -> CountDocuments<'_>

Gets the number of documents.

Note that this method returns an accurate count.

await will return Result<u64>.

These methods can be chained before .await to set options:

session

with_options

hint

limit

max_time

skip

collation

selection_criteria

read_concern

comment

impl<T> Collection<T>
where T: Send + Sync,

pub fn create_index(&self, index: IndexModel) -> CreateIndex<'_>

Creates the given index on this collection.

await will return Result<CreateIndexResult>.

These methods can be chained before .await to set options:

session

with_options

commit_quorum

max_time

write_concern

comment

pub fn create_indexes( &self, indexes: impl IntoIterator<Item = IndexModel>, ) -> CreateIndex<'_, Multiple>

Creates the given indexes on this collection.

await will return Result<CreateIndexesResult>.

These methods can be chained before .await to set options:

session

with_options

commit_quorum

max_time

write_concern

comment

impl<T> Collection<T>
where T: Send + Sync,

pub fn delete_one(&self, query: Document) -> Delete<'_>

Deletes up to one document found matching query.

This operation will retry once upon failure if the connection and encountered error support retryability. See the documentation here for more information on retryable writes.

await will return Result<DeleteResult>.

These methods can be chained before .await to set options:

session

with_options

collation

write_concern

hint

let_vars

comment

pub fn delete_many(&self, query: Document) -> Delete<'_>

Deletes all documents stored in the collection matching query.

await will return Result<DeleteResult>.

These methods can be chained before .await to set options:

session

with_options

collation

write_concern

hint

let_vars

comment

impl<T> Collection<T>
where T: Send + Sync,

pub fn distinct( &self, field_name: impl AsRef<str>, filter: Document, ) -> Distinct<'_>

Finds the distinct values of the field specified by field_name across the collection.

await will return Result<Vec<Bson>>.

These methods can be chained before .await to set options:

session

with_options

max_time

selection_criteria

read_concern

collation

comment

impl<T> Collection<T>
where T: Send + Sync,

pub fn drop(&self) -> DropCollection<'_>

Drops the collection, deleting all data and indexes stored in it.

await will return Result<()>.

These methods can be chained before .await to set options:

session

with_options

write_concern

encrypted_fields

impl<T> Collection<T>
where T: Send + Sync,

pub fn drop_index(&self, name: impl AsRef<str>) -> DropIndex<'_>

Drops the index specified by name from this collection.

await will return Result<()>.

These methods can be chained before .await to set options:

session

with_options

max_time

write_concern

comment

pub fn drop_indexes(&self) -> DropIndex<'_>

Drops all indexes associated with this collection.

await will return Result<()>.

These methods can be chained before .await to set options:

session

with_options

max_time

write_concern

comment

impl<T: Send + Sync> Collection<T>

pub fn find(&self, filter: Document) -> Find<'_, T>

Finds the documents in the collection matching filter.

await will return Result<Cursor<T>> (or Result<SessionCursor<T>> if a session is provided).

These methods can be chained before .await to set options:

session

with_options

allow_disk_use

allow_partial_results

batch_size

comment

cursor_type

hint

limit

max

max_await_time

max_scan

max_time

min

no_cursor_timeout

projection

read_concern

return_key

selection_criteria

show_record_id

skip

sort

collation

let_vars

impl<T: DeserializeOwned + Send + Sync> Collection<T>

pub fn find_one(&self, filter: Document) -> FindOne<'_, T>

Finds a single document in the collection matching filter.

await will return Result<Option<T>>.

These methods can be chained before .await to set options:

session

with_options

allow_partial_results

collation

comment

hint

max

max_scan

max_time

min

projection

read_concern

return_key

selection_criteria

show_record_id

skip

sort

let_vars

impl<T: DeserializeOwned + Send + Sync> Collection<T>

pub fn find_one_and_delete(&self, filter: Document) -> FindOneAndDelete<'_, T>

Atomically finds up to one document in the collection matching filter and deletes it.

This operation will retry once upon failure if the connection and encountered error support retryability. See the documentation here for more information on retryable writes.

await will return Result<Option<T>>.

These methods can be chained before .await to set options:

session

with_options

max_time

projection

sort

write_concern

collation

hint

let_vars

comment

pub fn find_one_and_update( &self, filter: Document, update: impl Into<UpdateModifications>, ) -> FindOneAndUpdate<'_, T>

Atomically finds up to one document in the collection matching filter and updates it. Both Document and Vec<Document> implement Into<UpdateModifications>, so either can be passed in place of constructing the enum case. Note: pipeline updates are only supported in MongoDB 4.2+.

This operation will retry once upon failure if the connection and encountered error support retryability. See the documentation here for more information on retryable writes.

await will return Result<Option<T>>.

These methods can be chained before .await to set options:

session

with_options

array_filters

bypass_document_validation

max_time

projection

return_document

sort

upsert

write_concern

collation

hint

let_vars

comment

impl<T: Serialize + DeserializeOwned + Send + Sync> Collection<T>

pub fn find_one_and_replace( &self, filter: Document, replacement: impl Borrow<T>, ) -> FindOneAndReplace<'_, T>

Atomically finds up to one document in the collection matching filter and replaces it with replacement.

This operation will retry once upon failure if the connection and encountered error support retryability. See the documentation here for more information on retryable writes.

await will return Result<Option<T>>.

These methods can be chained before .await to set options:

session

with_options

bypass_document_validation

max_time

projection

return_document

sort

upsert

write_concern

collation

hint

let_vars

comment

impl<T: Serialize + Send + Sync> Collection<T>

pub fn insert_many( &self, docs: impl IntoIterator<Item = impl Borrow<T>>, ) -> InsertMany<'_>

Inserts the data in docs into the collection.

Note that this method accepts both owned and borrowed values, so the input documents do not need to be cloned in order to be passed in.

This operation will retry once upon failure if the connection and encountered error support retryability. See the documentation here for more information on retryable writes.

await will return Result<InsertManyResult>.

These methods can be chained before .await to set options:

session

with_options

bypass_document_validation

ordered

write_concern

comment

impl<T: Serialize + Send + Sync> Collection<T>

pub fn insert_one(&self, doc: impl Borrow<T>) -> InsertOne<'_>

Inserts doc into the collection.

Note that either an owned or borrowed value can be inserted here, so the input document does not need to be cloned to be passed in.

This operation will retry once upon failure if the connection and encountered error support retryability. See the documentation here for more information on retryable writes.

await will return Result<InsertOneResult>.

These methods can be chained before .await to set options:

session

with_options

bypass_document_validation

write_concern

comment

impl<T> Collection<T>
where T: Send + Sync,

pub fn list_indexes(&self) -> ListIndexes<'_>

Lists all indexes on this collection.

await will return Result<Cursor<IndexModel>> (or Result<SessionCursor<IndexModel>> if a ClientSession is provided).

These methods can be chained before .await to set options:

with_options

max_time

batch_size

comment

session

pub fn list_index_names(&self) -> ListIndexes<'_, ListNames>

Gets the names of all indexes on the collection.

await will return Result<Vec<String>>.

These methods can be chained before .await to set options:

with_options

max_time

batch_size

comment

session

impl<T: Serialize + Send + Sync> Collection<T>

pub fn replace_one( &self, query: Document, replacement: impl Borrow<T>, ) -> ReplaceOne<'_>

Replaces up to one document matching query in the collection with replacement.

This operation will retry once upon failure if the connection and encountered error support retryability. See the documentation here for more information on retryable writes.

await will return Result<UpdateResult>.

These methods can be chained before .await to set options:

session

with_options

bypass_document_validation

upsert

collation

hint

write_concern

let_vars

comment

sort

impl<T> Collection<T>
where T: Send + Sync,

pub fn create_search_indexes( &self, models: impl IntoIterator<Item = SearchIndexModel>, ) -> CreateSearchIndex<'_, Multiple>

Creates multiple search indexes on the collection.

await will return Result<Vec<String>>.

These methods can be chained before .await to set options:

with_options

pub fn create_search_index( &self, model: SearchIndexModel, ) -> CreateSearchIndex<'_, Single>

Convenience method for creating a single search index.

await will return Result<String>.

These methods can be chained before .await to set options:

with_options

pub fn update_search_index( &self, name: impl Into<String>, definition: Document, ) -> UpdateSearchIndex<'_>

Updates the search index with the given name to use the provided definition.

await will return [Result<()>].

These methods can be chained before .await to set options:

with_options

pub fn drop_search_index(&self, name: impl Into<String>) -> DropSearchIndex<'_>

Drops the search index with the given name.

await will return [Result<()>].

These methods can be chained before .await to set options:

with_options

pub fn list_search_indexes(&self) -> ListSearchIndexes<'_>

Gets index information for one or more search indexes in the collection.

If name is not specified, information for all indexes on the specified collection will be returned.

await will return Result<Cursor<Document>>.

These methods can be chained before .await to set options:

name

aggregate_options

with_options

impl<T> Collection<T>
where T: Send + Sync,

pub fn update_many( &self, query: Document, update: impl Into<UpdateModifications>, ) -> Update<'_>

Updates all documents matching query in the collection.

Both Document and Vec<Document> implement Into<UpdateModifications>, so either can be passed in place of constructing the enum case. Note: pipeline updates are only supported in MongoDB 4.2+. See the official MongoDB documentation for more information on specifying updates.

await will return Result<UpdateResult>.

These methods can be chained before .await to set options:

session

with_options

array_filters

bypass_document_validation

upsert

collation

hint

write_concern

let_vars

comment

sort

pub fn update_one( &self, query: Document, update: impl Into<UpdateModifications>, ) -> Update<'_>

Updates up to one document matching query in the collection.

Both Document and Vec<Document> implement Into<UpdateModifications>, so either can be passed in place of constructing the enum case. Note: pipeline updates are only supported in MongoDB 4.2+. See the official MongoDB documentation for more information on specifying updates.

This operation will retry once upon failure if the connection and encountered error support retryability. See the documentation here for more information on retryable writes.

await will return Result<UpdateResult>.

These methods can be chained before .await to set options:

session

with_options

array_filters

bypass_document_validation

upsert

collation

hint

write_concern

let_vars

comment

sort

impl<T> Collection<T>
where T: Send + Sync,

pub fn watch(&self) -> Watch<'_, T>

Starts a new ChangeStream that receives events for all changes in this collection. A ChangeStream cannot be started on system collections.

See the documentation here on change streams.

Change streams require either a “majority” read concern or no read concern. Anything else will cause a server error.

await will return Result<ChangeStream<ChangeStreamEvent<T>>> or Result<SessionChangeStream<ChangeStreamEvent<T>>> if a ClientSession has been provided.

These methods can be chained before .await to set options:

pipeline

resume_after

with_options

full_document

full_document_before_change

start_at_operation_time

start_after

max_await_time

batch_size

collation

read_concern

selection_criteria

show_expanded_events

comment

session

impl<T> Collection<T>
where T: Send + Sync + Serialize,

pub fn insert_one_model( &self, document: impl Borrow<T>, ) -> Result<InsertOneModel>

Constructs an InsertOneModel with this collection’s namespace by serializing the provided value into a Document. Returns an error if serialization fails.

Note that the returned value must be provided to bulk_write for the insert to be performed.

pub fn replace_one_model( &self, filter: Document, replacement: impl Borrow<T>, ) -> Result<ReplaceOneModel>

Constructs a ReplaceOneModel with this collection’s namespace by serializing the provided value into a Document. Returns an error if serialization fails.

Note that the returned value must be provided to bulk_write for the replace to be performed.

impl<T> Collection<T>
where T: Send + Sync,

pub fn clone_with_type<U: Send + Sync>(&self) -> Collection<U>

Gets a clone of the Collection with a different type U.

pub fn client(&self) -> &Client

Get the Client that this collection descended from.

pub fn name(&self) -> &str

Gets the name of the Collection.

pub fn namespace(&self) -> Namespace

Gets the namespace of the Collection.

The namespace of a MongoDB collection is the concatenation of the name of the database containing it, the ‘.’ character, and the name of the collection itself. For example, if a collection named “bar” is created in a database named “foo”, the namespace of the collection is “foo.bar”.

pub fn selection_criteria(&self) -> Option<&SelectionCriteria>

Gets the selection criteria of the Collection.

pub fn read_concern(&self) -> Option<&ReadConcern>

Gets the read concern of the Collection.

pub fn write_concern(&self) -> Option<&WriteConcern>

Gets the write concern of the Collection.