Module mongodb::topology

MongoDB server set topology and asynchronous monitoring.

Modules

monitor
Asynchronous server and topology discovery and monitoring using isMaster results.

server
MongoDB server representation.

Structs

Topology
Holds status and connection information about a server set.

TopologyDescription
Topology information gathered from server set monitoring.

Enums

TopologyType
Describes the type of topology for a server set.

Constants

DEFAULT_HEARTBEAT_FREQUENCY_MS

DEFAULT_LOCAL_THRESHOLD_MS

DEFAULT_SERVER_SELECTION_TIMEOUT_MS