Struct mongodb::topology::monitor::IsMasterResult

pub struct IsMasterResult {
    pub ok: bool,
    pub is_master: bool,
    pub max_bson_object_size: i64,
    pub max_message_size_bytes: i64,
    pub local_time: Option<DateTime<UTC>>,
    pub min_wire_version: i64,
    pub max_wire_version: i64,
    pub msg: String,
    pub is_replica_set: bool,
    pub is_secondary: bool,
    pub me: Option<Host>,
    pub hosts: Vec<Host>,
    pub passives: Vec<Host>,
    pub arbiters: Vec<Host>,
    pub arbiter_only: bool,
    pub tags: BTreeMap<String, String>,
    pub set_name: String,
    pub election_id: Option<ObjectId>,
    pub primary: Option<Host>,
    pub hidden: bool,
    pub set_version: Option<i64>,
}

The result of an isMaster operation.

Fields
ok: bool is_master: bool max_bson_object_size: i64 max_message_size_bytes: i64 local_time: Option<DateTime<UTC>> min_wire_version: i64 max_wire_version: i64 msg: String

Shard-specific. mongos instances will add this field to the isMaster reply, and it will contain the value "isdbgrid".

is_replica_set: bool is_secondary: bool me: Option<Host> hosts: Vec<Host> passives: Vec<Host> arbiters: Vec<Host> arbiter_only: bool tags: BTreeMap<String, String> set_name: String election_id: Option<ObjectId> primary: Option<Host> hidden: bool set_version: Option<i64>
Methods

impl IsMasterResult

fn new(doc: Document) -> Result<IsMasterResult>

Parses an isMaster response document from the server.