Struct mongodb::topology::server::Server

pub struct Server {
    pub host: Host,
    pub description: Arc<RwLock<ServerDescription>>,
    // some fields omitted
}

Holds status and connection information about a single server.

Fields

host

Host connection details.

description

Monitored server information.

Methods

impl Server

fn new(client: Client, host: Host, top_description: Arc<RwLock<TopologyDescription>>, run_monitor: bool) -> Server

Returns a new server with the given host, initializing a new connection pool and monitor.

fn acquire_stream(&self) -> Result<PooledStream>

Returns a server stream from the connection pool.

fn request_update(&self)

Request an update from the monitor on the server status.