Struct mongodb::wire_protocol::flags::OpQueryFlags

pub struct OpQueryFlags {
    pub tailable_cursor: bool,
    pub slave_ok: bool,
    pub oplog_relay: bool,
    pub no_cursor_timeout: bool,
    pub await_data: bool,
    pub exhaust: bool,
    pub partial: bool,
}

Represents the bit vector of flags for an OP_QUERY message.

Fields
tailable_cursor: bool slave_ok: bool oplog_relay: bool no_cursor_timeout: bool await_data: bool exhaust: bool partial: bool
Methods

impl OpQueryFlags

fn no_flags() -> OpQueryFlags

Constructs a new struct with all flags set to false.

Return value

Returns the newly-created struct.

fn with_find_options(options: &FindOptions) -> OpQueryFlags

Constructs a new struct with flags based on a FindOptions struct.

Arguments

options - Struct whose fields contain the flags to initialize the new OpQueryFlags with

Return value

Returns the newly created OpQueryFlags struct.

fn to_i32(&self) -> i32

Gets the actual bit vector that the struct represents.

Return value

Returns the bit vector as an i32.