Struct mongodb::wire_protocol::flags::OpUpdateFlags

pub struct OpUpdateFlags {
    pub upsert: bool,
    pub multi_update: bool,
}

Represents the bit vector of options for an OP_UPDATE message.

Fields
upsert: bool multi_update: bool
Methods

impl OpUpdateFlags

fn no_flags() -> OpUpdateFlags

Constructs a new struct with all flags set to false.

Return value

Returns the newly-created struct.

fn to_i32(&self) -> i32

Gets the actual bit vector that the struct represents.

Return value

Returns the bit vector as an i32.