Module mongodb::action::gridfs

Action builders for gridfs.

Structs

Delete

Deletes a specific [FilesCollectionDocument] and its associated chunks. Construct with GridFsBucket::delete.

Drop

Removes all of the files and their associated chunks from a bucket. Construct with GridFsBucket::drop.

Find

Finds and returns the FilesCollectionDocuments within a bucket that match a given filter. Construct with GridFsBucket::find.

FindOne

Finds and returns a single FilesCollectionDocument within a bucket that matches a given filter. Construct with GridFsBucket::find_one.

OpenDownloadStream

Opens and returns a GridFsDownloadStream from which the application can read the contents of the stored file specified by an id. Construct with GridFsBucket::open_download_stream.

OpenDownloadStreamByName

Opens and returns a GridFsDownloadStream from which the application can read the contents of the stored file specified by a filename. Construct with GridFsBucket::open_download_stream_by_name.

OpenUploadStream

Creates and returns a GridFsUploadStream that the application can write the contents of a file to. Construct with GridFsBucket::open_upload_stream.

Rename

Renames a file. Construct with GridFsBucket::rename.