Module mongodb::change_stream::event

Contains the types related to a ChangeStream event.

Structs

ChangeNamespace

Identifies the collection or database on which an event occurred.

ChangeStreamEvent

A ChangeStreamEvent represents a change event in the associated change stream.

ResumeToken

An opaque token used for resuming an interrupted ChangeStream.

TruncatedArray

Describes an array that has been truncated.

UpdateDescription

Describes which fields have been updated or removed from a document.

Enums

OperationType

The operation type represented in a given change notification.