Module mongodb::coll::options

Options for collection-level operations.

Structs

AggregateOptions
Options for aggregation queries.

CountOptions
Options for count queries.

DistinctOptions
Options for distinct queries.

FindOneAndDeleteOptions
Options for findOneAndDelete operations.

FindOneAndUpdateOptions
Options for findOneAndUpdate operations.

FindOptions
Options for collection queries.

IndexModel
A single index model.

IndexOptions
Options for index operations.

InsertManyOptions
Options for insertMany operations.

UpdateOptions
Options for update operations.

Enums

CursorType
Describes the type of cursor to return on collection queries.

ReturnDocument
Describes the type of document to return on write operations.

WriteModel
Marker interface for writes that can be batched together.

Type Definitions

ReplaceOptions