Module mongodb::error

Contains the Error and Result types that mongodb uses.

Structs

BulkWriteError
An error that occurred during a write operation consisting of multiple writes that wasn't due to being unable to satisfy a write concern.

BulkWriteFailure
The set of errors that occurred during a write operation.

CommandError
An error that occurred due to a database command failing.

Error
An error that can occur in the mongodb crate. The inner ErrorKind is wrapped in an Arc to allow the errors to be cloned.

WriteConcernError
An error that occurred due to not being able to satisfy a write concern.

WriteError
An error that occurred during a write operation that wasn't due to being unable to satisfy a write concern.

Enums

ErrorKind
The types of errors that can occur.

WriteFailure
An error that occurred when trying to execute a write operation.

Type Definitions

Result
The result type for all methods that can return an error in the mongodb crate.