Enum mongodb::error::Error

pub enum Error {
    IoError(Error),
    EncoderError(EncoderError),
    DecoderError(DecoderError),
    OIDError(Error),
    FromHexError(FromHexError),
    WriteError(WriteException),
    BulkWriteError(BulkWriteException),
    ArgumentError(String),
    OperationError(String),
    ResponseError(String),
    CursorNotFoundError,
    PoisonLockError,
    CodedError(ErrorCode),
    EventListenerError(Option<Box<Error>>),
    MaliciousServerError(MaliciousServerErrorType),
    DefaultError(String),
}

The error type for MongoDB operations.

Variants

IoError

I/O operation errors of Read, Write, Seek, and associated traits.

EncoderError

A BSON struct could not be encoded.

DecoderError

A BSON struct could not be decoded.

OIDError

An ObjectId could not be generated.

FromHexError

A hexadecimal string could not be converted to bytes.

WriteError

A single-write operation failed.

BulkWriteError

A bulk-write operation failed due to one or more lower-level write-related errors.

ArgumentError

An invalid function or operational argument was provided.

OperationError

A database operation failed to send or receive a reply.

ResponseError

A database operation returned an invalid reply.

CursorNotFoundError

A cursor operation failed to return a cursor.

PoisonLockError

The application failed to secure a mutex due to a poisoned lock.

CodedError

A server error with a given code.

EventListenerError

The client was unable to emit the events to the listeners due to a poisoned lock; all event listeners were dropped, so they will have to be registered again. If the client is unable to emit a failure result, the error it failed to report is bundled into the EventListenerError.

MaliciousServerError

The server that the client is attempting to authenticate to does not actually have the user's authentication information stored.

DefaultError

A standard error with a string description; a more specific error should generally be used.