Module mongodb::event::sdam

Contains the events and functionality for monitoring Server Discovery and Monitoring.

Structs

ServerClosedEvent

Published when a server is closed.

ServerDescriptionChangedEvent

Published when a server description changes.

ServerHeartbeatFailedEvent

Published when a server monitor’s hello or legacy hello command fails.

ServerHeartbeatStartedEvent

Published when a server monitor’s hello or legacy hello command is started.

ServerHeartbeatSucceededEvent

Published when a server monitor’s hello or legacy hello command succeeds.

ServerOpeningEvent

Published when a server is initialized.

TopologyClosedEvent

Published when a topology is closed. Note that this event will not be published until the client associated with the topology is dropped.

TopologyDescription

A description of the most up-to-date information known about a topology. Further details can be found in the Server Discovery and Monitoring specification.

TopologyDescriptionChangedEvent

Published when a topology description changes.

TopologyOpeningEvent

Published when a topology is initialized.

Enums

SdamEvent

TopologyType

The possible types for a topology.

Traits

SdamEventHandlerDeprecated

Usage of this trait is deprecated. Applications should use the EventHandler API.

Type Aliases

ServerDescription

A description of the most up-to-date information known about a server. Further details can be found in the Server Discovery and Monitoring specification.