Struct mongodb::options::AggregateOptions

#[non_exhaustive]
pub struct AggregateOptions {
    pub allow_disk_use: Option<bool>,
    pub batch_size: Option<u32>,
    pub bypass_document_validation: Option<bool>,
    pub collation: Option<Collation>,
    pub comment: Option<String>,
    pub hint: Option<Hint>,
    pub max_await_time: Option<Duration>,
    pub max_time: Option<Duration>,
    pub read_concern: Option<ReadConcern>,
    pub selection_criteria: Option<SelectionCriteria>,
    pub write_concern: Option<WriteConcern>,
    pub let_vars: Option<Document>,
}

Specifies the options to a Collection::aggregate operation.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

allow_disk_use: Option<bool>

Enables writing to temporary files. When set to true, aggregation stages can write data to the _tmp subdirectory in the dbPath directory.

batch_size: Option<u32>

The number of documents the server should return per cursor batch.

Note that this does not have any affect on the documents that are returned by a cursor, only the number of documents kept in memory at a given time (and by extension, the number of round trips needed to return the entire set of documents returned by the query).

bypass_document_validation: Option<bool>

Opt out of document-level validation.

collation: Option<Collation>

The collation to use for the operation.

See the documentation for more information on how to use this option.

comment: Option<String>

Tags the query with an arbitrary string to help trace the operation through the database profiler, currentOp and logs.

hint: Option<Hint>

The index to use for the operation.

max_await_time: Option<Duration>

The maximum amount of time for the server to wait on new documents to satisfy a tailable await cursor query.

This option will have no effect on non-tailable cursors that result from this operation.

max_time: Option<Duration>

The maximum amount of time to allow the query to run.

This options maps to the maxTimeMS MongoDB query option, so the duration will be sent across the wire as an integer number of milliseconds.

read_concern: Option<ReadConcern>

The read concern to use for the operation.

If none is specified, the read concern defined on the object executing this operation will be used.

selection_criteria: Option<SelectionCriteria>

The criteria used to select a server for this operation.

If none is specified, the selection criteria defined on the object executing this operation will be used.

write_concern: Option<WriteConcern>

The write concern to use for the operation.

If none is specified, the write concern defined on the object executing this operation will be used.

let_vars: Option<Document>

A document with any amount of parameter names, each followed by definitions of constants in the MQL Aggregate Expression language. Each parameter name is then usable to access the value of the corresponding MQL Expression with the “$$” syntax within Aggregate Expression contexts.

This feature is only available on server versions 5.0 and above.

Implementations

impl AggregateOptions

pub fn builder(
) -> AggregateOptionsBuilder<((), (), (), (), (), (), (), (), (), (), (), ())>

Create a builder for building AggregateOptions. On the builder, call .allow_disk_use(...)(optional), .batch_size(...)(optional), .bypass_document_validation(...)(optional), .collation(...)(optional), .comment(...)(optional), .hint(...)(optional), .max_await_time(...)(optional), .max_time(...)(optional), .read_concern(...)(optional), .selection_criteria(...)(optional), .write_concern(...)(optional), .let_vars(...)(optional) to set the values of the fields. Finally, call .build() to create the instance of AggregateOptions.