Module mongodb::results

Contains the types of results returned by CRUD operations.

Structs

CollectionSpecification

Information about a collection as reported by Database::list_collections.

CollectionSpecificationInfo

Info about the collection that is contained in the CollectionSpecification::info field of a specification returned from Database::list_collections.

CreateIndexResult

Information about the index created as a result of a Collection::create_index.

CreateIndexesResult

Information about the indexes created as a result of a Collection::create_indexes.

DatabaseSpecification

A struct modeling the information about an individual database returned from Client::list_databases.

DeleteResult

The result of a Collection::delete_one or Collection::delete_many operation.

InsertManyResult

The result of a Collection::insert_many operation.

InsertOneResult

The result of a Collection::insert_one operation.

UpdateResult

The result of a Collection::update_one or Collection::update_many operation.

Enums

CollectionType

Describes the type of data store returned when executing Database::list_collections.