Struct mongodb::Collection

pub struct Collection { /* fields omitted */ }

Collection is the client-side abstraction of a MongoDB Collection. It can be used to perform collection-level operations such as CRUD operations. A Collection can be obtained through a Database by calling either Database::collection or Database::collection_with_options.

Collection uses std::sync::Arc internally, so it can safely be shared across threads. For example:

let coll = client.database("items").collection("in_stock");

for i in 0..5 {
    let coll_ref = coll.clone();

    std::thread::spawn(move || {
        // Perform operations with `coll_ref`. For example:
        coll_ref.insert_one(doc! { "x": i }, None);
    });
}

Methods

impl Collection

pub fn name(&self) -> &str

Gets the name of the Collection.

pub fn namespace(&self) -> Namespace

Gets the namespace of the Collection.

The namespace of a MongoDB collection is the concatenation of the name of the database containing it, the '.' character, and the name of the collection itself. For example, if a collection named "bar" is created in a database named "foo", the namespace of the collection is "foo.bar".

pub fn selection_criteria(&self) -> Option<&SelectionCriteria>

Gets the selection criteria of the Collection.

pub fn read_concern(&self) -> Option<&ReadConcern>

Gets the read concern of the Collection.

pub fn write_concern(&self) -> Option<&WriteConcern>

Gets the write concern of the Collection.

pub fn drop(
&self,
options: impl Into<Option<DropCollectionOptions>>
) -> Result<()>

Drops the collection, deleting all data, users, and indexes stored in it.

pub fn aggregate(
&self,
pipeline: impl IntoIterator<Item = Document>,
options: impl Into<Option<AggregateOptions>>
) -> Result<Cursor>

Runs an aggregation operation.

See the documentation here for more information on aggregations.

pub fn estimated_document_count(
&self,
options: impl Into<Option<EstimatedDocumentCountOptions>>
) -> Result<i64>

Estimates the number of documents in the collection using collection metadata.

pub fn count_documents(
&self,
filter: impl Into<Option<Document>>,
options: impl Into<Option<CountOptions>>
) -> Result<i64>

Gets the number of documents matching filter.

Note that using Collection::estimated_document_count is recommended instead of this method is most cases.

pub fn delete_many(
&self,
query: Document,
options: impl Into<Option<DeleteOptions>>
) -> Result<DeleteResult>

Deletes all documents stored in the collection matching query.

pub fn delete_one(
&self,
query: Document,
options: impl Into<Option<DeleteOptions>>
) -> Result<DeleteResult>

Deletes up to one document found matching query.

pub fn distinct(
&self,
field_name: &str,
filter: impl Into<Option<Document>>,
options: impl Into<Option<DistinctOptions>>
) -> Result<Vec<Bson>>

Finds the distinct values of the field specified by field_name across the collection.

pub fn find(
&self,
filter: impl Into<Option<Document>>,
options: impl Into<Option<FindOptions>>
) -> Result<Cursor>

Finds the documents in the collection matching filter.

pub fn find_one(
&self,
filter: impl Into<Option<Document>>,
options: impl Into<Option<FindOneOptions>>
) -> Result<Option<Document>>

Finds a single document in the collection matching filter.

pub fn find_one_and_delete(
&self,
filter: Document,
options: impl Into<Option<FindOneAndDeleteOptions>>
) -> Result<Option<Document>>

Atomically finds up to one document in the collection matching filter and deletes it.

pub fn find_one_and_replace(
&self,
filter: Document,
replacement: Document,
options: impl Into<Option<FindOneAndReplaceOptions>>
) -> Result<Option<Document>>

Atomically finds up to one document in the collection matching filter and replaces it with replacement.

pub fn find_one_and_update(
&self,
filter: Document,
update: impl Into<UpdateModifications>,
options: impl Into<Option<FindOneAndUpdateOptions>>
) -> Result<Option<Document>>

Atomically finds up to one document in the collection matching filter and updates it. Both Document and Vec<Document> implement Into<UpdateModifications>, so either can be passed in place of constructing the enum case. Note: pipeline updates are only supported in MongoDB 4.2+.

pub fn insert_many(
&self,
docs: impl IntoIterator<Item = Document>,
options: impl Into<Option<InsertManyOptions>>
) -> Result<InsertManyResult>

Inserts the documents in docs into the collection.

pub fn insert_one(
&self,
doc: Document,
options: impl Into<Option<InsertOneOptions>>
) -> Result<InsertOneResult>

Inserts doc into the collection.

pub fn replace_one(
&self,
query: Document,
replacement: Document,
options: impl Into<Option<ReplaceOptions>>
) -> Result<UpdateResult>

Replaces up to one document matching query in the collection with replacement.

pub fn update_many(
&self,
query: Document,
update: impl Into<UpdateModifications>,
options: impl Into<Option<UpdateOptions>>
) -> Result<UpdateResult>

Updates all documents matching query in the collection.

Both Document and Vec<Document> implement Into<UpdateModifications>, so either can be passed in place of constructing the enum case. Note: pipeline updates are only supported in MongoDB 4.2+. See the official MongoDB documentation for more information on specifying updates.

pub fn update_one(
&self,
query: Document,
update: impl Into<UpdateModifications>,
options: impl Into<Option<UpdateOptions>>
) -> Result<UpdateResult>

Updates up to one document matching query in the collection.

Both Document and Vec<Document> implement Into<UpdateModifications>, so either can be passed in place of constructing the enum case. Note: pipeline updates are only supported in MongoDB 4.2+. See the official MongoDB documentation for more information on specifying updates.

pub fn create_indexes(
&self,
models: impl IntoIterator<Item = IndexModel>
) -> Result<Vec<String>>

Creates the indexes specified by models.

pub fn drop_index(&self, name: &str) -> Result<Document>

Drops the index specified by name.

pub fn drop_index_with_keys(&self, keys: Document) -> Result<Document>

Drops the index with the given keys.

pub fn drop_indexes(&self) -> Result<Document>

Drops all indexes in the collection.