Struct mongodb::topology::monitor::IsMasterResult

pub struct IsMasterResult {
    pub ok: bool,
    pub is_master: bool,
    pub max_bson_object_size: i64,
    pub max_message_size_bytes: i64,
    pub local_time: Option<DateTime<UTC>>,
    pub min_wire_version: i64,
    pub max_wire_version: i64,
    pub msg: String,
    pub is_replica_set: bool,
    pub is_secondary: bool,
    pub me: Option<Host>,
    pub hosts: Vec<Host>,
    pub passives: Vec<Host>,
    pub arbiters: Vec<Host>,
    pub arbiter_only: bool,
    pub tags: BTreeMap<String, String>,
    pub set_name: String,
    pub election_id: Option<ObjectId>,
    pub primary: Option<Host>,
    pub hidden: bool,
}

The result of an isMaster operation.

Fields

ok

is_master

max_bson_object_size

max_message_size_bytes

local_time

min_wire_version

max_wire_version

msg

Shard-specific. mongos instances will add this field to the isMaster reply, and it will contain the value "isdbgrid".

is_replica_set

is_secondary

me

hosts

passives

arbiters

arbiter_only

tags

set_name

election_id

primary

hidden

Methods

impl IsMasterResult

fn new(doc: Document) -> Result<IsMasterResult>

Parses an isMaster response document from the server.