Struct mongodb::topology::server::ServerDescription

pub struct ServerDescription {
    pub server_type: ServerType,
    pub err: Arc<Option<Error>>,
    pub round_trip_time: Option<i64>,
    pub min_wire_version: i64,
    pub max_wire_version: i64,
    pub me: Option<Host>,
    pub hosts: Vec<Host>,
    pub passives: Vec<Host>,
    pub arbiters: Vec<Host>,
    pub tags: BTreeMap<String, String>,
    pub set_name: String,
    pub election_id: Option<ObjectId>,
    pub primary: Option<Host>,
}

Server information gathered from server monitoring.

Fields

server_type

The server type.

err

Any error encountered while monitoring this server.

round_trip_time

The average round-trip time over the last 5 monitoring checks.

min_wire_version

The minimum wire version supported by this server.

max_wire_version

The maximum wire version supported by this server.

me

The server's host information, if it is part of a replica set.

hosts

All hosts in the replica set known by this server.

passives

All passive members of the replica set known by this server.

arbiters

All arbiters in the replica set known by this server.

tags

Server tags for targeted read operations on specific replica set members.

set_name

The replica set name.

election_id

The server's current election id, if it believes it is a primary.

primary

The server's opinion of who the primary is.

Methods

impl ServerDescription

fn new() -> ServerDescription

Returns a default, unknown server description.

fn update(&mut self, ismaster: IsMasterResult, round_trip_time: i64)

fn set_err(&mut self, err: Error)

fn clear(&mut self)